remove confluent from the consumer.
